4.21*: Blog List Expansion, step 9 submission

+++ part4/controllers/blogs.js
@@ -49,6 +49,18 @@ blogsRouter.put('/:id', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
+  const blog = await Blog.findById(req.params.id)
+
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!decodedToken) {
+    res.status(401).json({ error: 'invalid token' })
+  }
+
+  if (decodedToken.id !== blog.user.toString()) {
+    return res
+      .status(401)
+      .json('Unauthorized request, blog was created by someone else')
+  }
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
